Refactor of verbal flashcards backend
Functionality has been split into four modules, all with distinct responsibilities. The endpoint class now functions as a slimmer routing class with more general logic.
